Move CountingAccessFileHiveMetastore to Hive
Also, rename to CountingAccessHiveMetastore.
